+++ packages/product-feed-interface/CHANGELOG.md
@@ -5,6 +5,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+- [`StandardFeedItemInterface`](src/StandardFeedItemInterface.php) contains ID of its main category (@MattCzerner)
 
 ## [0.4.0] - 2017-09-25
 ### Added